"Updated database migrations to change data types for 'price' and 'creditos' columns, and modified a button in the pricing-coins.blade.php view to open a payment form in a new window."

+++ AI-Maker/database/migrations/2024_07_25_093512_create_facturas_table.php
@@ -12,9 +12,9 @@ class CreateFacturasTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('user_id');
             $table->timestamp('date')->useCurrent(); // Usa la fecha actual por defecto
-            $table->double('price');
+            $table->decimal('price',8,2);
             $table->string('description')->nullable(); // Se agrega la columna description como string
-            $table->integer('creditos')->unsigned(); // Campo para almacenar los créditos comprados
+            $table->decimal('creditos',8,2)->unsigned(); // Campo para almacenar los créditos comprados
             $table->timestamps();
             
             // Foreign key constraints
